+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Fixed link to contribution guide in [`commonContributingGuide`](docs/validations.md#commoncontributingguide).
 
+### Added
+
+- New utility [`structuredFileAddedLines`](docs/utilities.md#structuredfileaddedlines) to get and object with the added line numbers and its content.
+- New utility [`structuredFileAddedLineMatches`](docs/utilities.md#structuredfileaddedlinematches) to get the list of line numbers of the added lines matching a pattern.
+
 ## [1.10.0] - 2019-5-3
 
 ### Fixed